New array-based architecture.
There is no more methods, but template function are stored in array uniques per datatype and component.
So far the function pointer have type void(*)(...).
Users of a type must define template function, and a template class/struct with a static data member structname::table that is the
table of functions.

A type-cheking mechanism could be done, by storing the function std::type_info with each function pointer and by
auto-adjoining the std::type_info of parameters of an actor message.